pyproject.toml
[project]
name = "BJTU-Underwater-Visual-System"
version = "0.1.0"
description = "None"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lap>=0.5.12",
    "opencv-contrib-python==4.12.0.88",
    "opencv-python==4.12.0.88",
    "pyyaml>=6.0.3",
    "serial>=0.0.97",
    "torch>=2.5.1; platform_machine == 'AMD64'",
    "torchvision>=0.20.1; platform_machine == 'AMD64'",
    "tqdm>=4.67.1",
    "ultralytics==8.3.228",
]

[tool.uv]
environments = [
    "platform_machine == 'AMD64'",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "platform_machine == 'AMD64'" },   # Windows
]
torchvision = [
    { index = "pytorch-cu128", marker = "platform_machine == 'AMD64'" },   # Windows
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
